+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fix visibility conditions on tree cascade dropdown questions
+- Fixed the `Tree cascade Dropdown` field so that the subtree depth limit is enforced when loading children via AJAX
+- Fixed the `Tree cascade Dropdown` question to only show items from the configured custom dropdown instead of all custom dropdowns
+- Fixed `Tree cascade Dropdown` question showing items from all custom dropdowns instead of only items from the configured one
 
 ## [1.1.1] - 2026-05-27